move to utils package
